Add headings, unordered and ordered list
The website was previously bland and formatted incorrectly.

To solve this use the <h3> tag to make the description of the product stand out.
Use <ul> and <li> to create an unordered list for the ingrediants.
Use <ol> and <li> to create an ordered list for the steps of how to make
the product.